[network] convert to a proper ConnectivityManagerBuilder
